requirements.txt
# FastMCP
fastmcp

# Config management
python-dotenv
pydantic

# Async support
asyncio